Add Taskcluster support to split-corpus.sh
In addition to the usual zstd support, this allows for optionally archiving all of the resulting files together. This is necessary in Taskcluster, because we must know the names of the files we need from other tasks prior to them running, and we don't know how many files this script will produce until runtime as things stand now.