Implement `train-teacher` step
Besides the basics, this required the introduction of a new `ensemble` kind. `train-teacher` (as well as a few other steps) may be "ensembled" - which means multiple versions of the same task will run in parallel. The only distinction between the tasks is in the name. (Ultimately they fold back into a downstream task later.)