zstd support and a minor fix for `download-mono.sh`
I originally bypassed this script in Taskcluster - but have since realized that it does some important filtering. Besides the zstd support being added, this commit ensures that `shuf` is fed an integer from `bc` (the version of `shuf` used in Taskcluster seems to treat it like an overflow if it's given a float).